item controller 작업 및 dto 수정

+++ src/main/java/com/woong/projectmanager/service/ItemService.java
@@ -3,7 +3,9 @@ package com.woong.projectmanager.service;
 import com.woong.projectmanager.domain.Channel;
 import com.woong.projectmanager.domain.Item;
 import com.woong.projectmanager.domain.Users;
-import com.woong.projectmanager.dto.ItemDto;
+import com.woong.projectmanager.dto.request.ItemAddRequestDto;
+import com.woong.projectmanager.dto.response.ItemResponseDto;
+import com.woong.projectmanager.exception.AccessTokenFailedException;
 import com.woong.projectmanager.repository.ChannelRepository;
 import com.woong.projectmanager.repository.ItemRepository;
 import com.woong.projectmanager.repository.UsersRepository;
@@ -26,11 +28,11 @@ public class ItemService {
     private final ChannelRepository channelRepository;
 
     @Transactional
-    public Item createItem(ItemDto itemDto){
+    public ItemResponseDto createItem(ItemAddRequestDto itemAddRequestDto, String writerEmail){
 
-        Item item = itemDto.toEntity();
-        Users writer = usersRepository.findByEmail(itemDto.getWriterEmail()).orElseThrow();
-        Channel channel = channelRepository.findById(itemDto.getChannelId()).orElseThrow();
+        Item item = itemAddRequestDto.toEntity();
+        Users writer = usersRepository.findByEmail(writerEmail).orElseThrow();
+        Channel channel = channelRepository.findById(itemAddRequestDto.getChannelId()).orElseThrow();
 
         //item 작성자 설정
         item.setWriter(writer);
@@ -38,35 +40,68 @@ public class ItemService {
         //item 추가
         channel.addItem(item);
 
-        return itemRepository.save(item);
+        //
+        itemRepository.save(item);
 
+        return new ItemResponseDto(item);
     }
 
     @Transactional
-    public void removeItem(Long itemId){
+    public ItemResponseDto updateItem(Long itemId,ItemAddRequestDto itemAddRequestDto, String requestEmail){
+
         Item item = itemRepository.findById(itemId).orElseThrow();
 
-        itemRepository.delete(item);
+        if(!item.getWriter().getEmail().equals(requestEmail)){
+            throw new AccessTokenFailedException(requestEmail);
+        }
+
+        item.update(itemAddRequestDto);
+
+        itemRepository.save(item);
+
+        ItemResponseDto itemResponseDto = new ItemResponseDto(item);
+
+        return itemResponseDto;
     }
 
-    public List<ItemDto> getItemList(Long channelId){
+    @Transactional
+    public ItemResponseDto removeItem(Long itemId,String requestEmail){
+
+        Item item = itemRepository.findById(itemId).orElseThrow();
+
+        if(!item.getWriter().getEmail().equals(requestEmail)){
+            throw new AccessTokenFailedException(requestEmail);
+        }
+
+        itemRepository.delete(item);
+
+        return new ItemResponseDto(item);
+    }
+
+    public List<ItemResponseDto> getItemList(Long channelId){
 
         Channel channel = channelRepository.findById(channelId).orElseThrow();
 
         //DTO 생성하여 반환
-        List<ItemDto> itemDtoList = new ArrayList<>();
+        List<ItemResponseDto> itemDtoList = new ArrayList<>();
 
         for(var el : channel.getItemList()){
-            ItemDto itemDto = new ItemDto();
+            ItemResponseDto itemResponseDto = new ItemResponseDto(el);
+            itemDtoList.add(itemResponseDto);
+        }
 
-            itemDto.setName(el.getName());
-            itemDto.setStatus(el.getStatus());
-            itemDto.setStartDate(el.getStartDate());
-            itemDto.setEndDate(el.getEndDate());
-            itemDto.setWriterEmail(el.getWriter().getEmail());
-            itemDto.setChannelId(el.getChannel().getId());
+        return itemDtoList;
 
-            itemDtoList.add(itemDto);
+    }
+
+    public List<ItemResponseDto> getItemList(Channel channel){
+
+        //DTO 생성하여 반환
+        List<ItemResponseDto> itemDtoList = new ArrayList<>();
+
+        for(var el : channel.getItemList()){
+            ItemResponseDto itemResponseDto = new ItemResponseDto(el);
+            itemDtoList.add(itemResponseDto);
         }
 
         return itemDtoList;
